Move the output extension registry to the main output module
